close #3
Previous commit called close #3 didn't work properly in testing, should
work properly now. Might add an additional line to the "There are
currently no listings." text that says "Make one today!" and acts as a
hyperlink, but it would just be an aesthetic thing and maybe not all
that useful or meaningful.